fix: eslint-config cypress and storybook definitions fixes
