+++ packages/eslint-config/src/overrides/storybook.ts
@@ -14,6 +14,10 @@ export const storybookOverrides: Linter.ConfigOverride = {
     'react-hooks/exhaustive-deps': 'off',
     'react/function-component-definition': 'off',
     'react/jsx-no-constructed-context-values': 'off',
+    'react/jsx-props-no-spreading': 'off',
+    'react/react-in-jsx-scope': 'off',
+    'react/prop-types': 'off',
+    'react/require-default-props': 'off',
     'react-refresh/only-export-components': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/naming-convention': 'off',
