+++ packages/eslint-config/src/presets/storybook.ts
@@ -1,19 +1,17 @@
 import type { Linter } from 'eslint';
 
-import { cypressPreset } from './cypressInternal';
+import { storybookOverrides } from '../overrides/storybook';
 
 if (!process.env.NODE_ENV) {
   // This check allows us to run linters inside IDE's
   process.env.NODE_ENV = 'development';
 }
 
-const { extends: extendPreset, ...config } = cypressPreset;
-
 module.exports = {
   overrides: [
     {
-      files: ['**/cypress/**/*', '**/*.{cy,cypress}.{js,jsx,ts,tsx}'],
-      ...config,
+      extends: ['@tablecheck/eslint-config/typescript'],
+      ...storybookOverrides,
     },
   ],
 } satisfies Linter.Config;
